tests: wget test code directly from github
... because downstream SHA1s are available only from github.

In the past, upstream was wgetting lots of stuff directly from github for their
automated integration testing. This turned out to be not so good, because
github would often time out or be unavailable.

To fix this, upstream started maintaining a mirror of the main github repo at
git.ceph.com.

While this mirror is accessible from anywhere, it does not have the downstream
SHA1s that we are testing, so the wget fails.

To fix this, revert to downloading directly from github (hoping that it will
work) and use SUSE/ceph.git instead of ceph/ceph.git.

Also, raw.github.com forwards to raw.githubusercontent, so avoid one redirect.

Fixes: https://bugzilla.suse.com/show_bug.cgi?id=1000105
Signed-off-by: Jan Fajerski <jfajerski@suse.com>
Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit f749ee89cae6b577005df1b5ab8166367133d20f)
Signed-off-by: Jan Fajerski <jfajerski@suse.com>