rpm: consider xio bcond on x86_64 and aarch64 only
We are building on s390x and ppc64le only for ceph-common and ceph-fuse, for
which libxio is not needed. The libxio package is not built for s390x and
ppc64le. We have to be heavy-handed because there is no way to deactivate the
Project Config macro in SES4.

Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit c1a27071ff8f69eee28c2f210040d515aaf953f3)

Conflicts:
	ceph.spec.in (trivial resolution)