build/ops: do_cmake.sh: get rid of which dependency
"which" is not available on minimal systems, whereas "type" is a shell builtin,
so the chances of the command "just working" are greater with type than with
which.

In other parts of Ceph build/ops we have replaced "which" with "type" to good
effect.

Signed-off-by: Nathan Cutler <ncutler@suse.com>
(cherry picked from commit a7173385582e73269ecc642fa3214a1165e1ba97)