[REMOVE] qa: deepsea: exclude salt packages
Currently we have a temporary situation where we have Salt 2017.7.2 installed
(and locked), so we fail on this line because it tries to install the latest
Salt 2018.1.99 and fails due to the lock.

This commit can be removed once DeepSea is fixed to work with the Salt that is
currently shipping with SLE15.

Signed-off-by: Nathan Cutler <ncutler@suse.com>