qa/tasks/deepsea.py: get it passing linter again
This file is supposed to be maintained in a linter-passing state.
This is defined in the comment at the top of the file:

"""
Task (and subtasks) for automating deployment of Ceph using DeepSea

Linter:
    flake8 --max-line-length=100
"""

Somehow it was not passing the linter -- either it started out that way,
or it drifted out of the ideal state over time.

Signed-off-by: Nathan Cutler <ncutler@suse.com>