qa/suites/deepsea: Switching to upgrade.status
Switching to upgrade.status instead of status.report as it shows more
clear information and it's in accordance with docs.

Also installing deepsea-cli on the master node in order to fix CI
failure due to failure of deepsea commands after upgrade.

suites/deepsea/tier4/upgrade:power-of-2 pg_num fix

Since pg_num being power of 2 is not a requirement for luminous but it
is on nautilus, disabling this warning on global.conf in order to fix
the test that fails to get HEALTH-OK status after upgrade.

Signed-off-by: Georgios Kyratsas <gkyratsas@suse.com>